fix bugs in the write-path affecting empty structs and List(Text)

+++ write_test.go
@@ -41,8 +41,8 @@ func ExampleAirplaneWrite() string {
 	// how to create a list. Requires a cast at the moment.
 	homes := air.NewAirportList(seg, 2)
 	uint16list := capn.UInt16List(homes) // cast to the underlying type
-	uint16list.Set(0, air.AIRPORT_JFK)
-	uint16list.Set(1, air.AIRPORT_LAX)
+	uint16list.Set(0, uint16(air.AIRPORT_JFK))
+	uint16list.Set(1, uint16(air.AIRPORT_LAX))
 
 	// set the primitive fields
 	planebase.SetCanFly(true)
